chore: rename fixtures
