+++ packages/data-loader/src/parsers/parseCsv/__tests__/parseCsv.test.ts
@@ -5,15 +5,15 @@ import {
 import fs from "fs";
 import path from "path";
 import { parseCsv } from "../index";
-const outputJson: Array<{
+const expectedJson: Array<{
   [k: string]: KintoneRecordField.OneOf;
-}> = require("./fixtures/output.json");
+}> = require("./fixtures/expected.json");
 const fieldsJson: {
   properties: { [k: string]: KintoneFormFieldProperty.OneOf };
 } = require("./fixtures/fields.json");
-const subTableOutputJson: Array<{
+const subTableExpectedJson: Array<{
   [k: string]: KintoneRecordField.OneOf;
-}> = require("./fixtures/subtable_output.json");
+}> = require("./fixtures/subtable_expected.json");
 const subTableFieldsJson: {
   properties: { [k: string]: KintoneFormFieldProperty.OneOf };
 } = require("./fixtures/subtable_fields.json");
@@ -24,13 +24,13 @@ describe("csvParser", () => {
       path.resolve(__dirname, "fixtures", "records.csv"),
       "utf-8"
     );
-    expect(parseCsv(csv, fieldsJson)).toEqual(outputJson);
+    expect(parseCsv(csv, fieldsJson)).toEqual(expectedJson);
   });
   it("should convert subtable included csv string to JSON correctly", async () => {
     const csv = await fs.promises.readFile(
       path.resolve(__dirname, "fixtures", "subtable_records.csv"),
       "utf-8"
     );
-    expect(parseCsv(csv, subTableFieldsJson)).toEqual(subTableOutputJson);
+    expect(parseCsv(csv, subTableFieldsJson)).toEqual(subTableExpectedJson);
   });
 });
