+++ packages/data-loader/src/printers/printAsCsv/__tests__/fixtures/fields.json
@@ -4,19 +4,19 @@
     "recordNumber": {
       "type": "RECORD_NUMBER",
       "code": "recordNumber",
-      "label": "レコード番号",
+      "label": "recordNumber",
       "noLabel": false
     },
     "updatedTime": {
       "type": "UPDATED_TIME",
       "code": "updatedTime",
-      "label": "更新日時",
+      "label": "updatedTime",
       "noLabel": false
     },
     "dropDown": {
       "type": "DROP_DOWN",
       "code": "dropDown",
-      "label": "ドロップダウン",
+      "label": "dropDown",
       "noLabel": false,
       "required": false,
       "options": {
@@ -34,31 +34,31 @@
     "creator": {
       "type": "CREATOR",
       "code": "creator",
-      "label": "作成者",
+      "label": "creator",
       "noLabel": false
     },
-    "作業者": {
+    "Assignee": {
       "type": "STATUS_ASSIGNEE",
-      "code": "作業者",
-      "label": "作業者",
+      "code": "Assignee",
+      "label": "Assignee",
       "enabled": false
     },
     "modifier": {
       "type": "MODIFIER",
       "code": "modifier",
-      "label": "更新者",
+      "label": "modifier",
       "noLabel": false
     },
-    "ステータス": {
+    "Status": {
       "type": "STATUS",
-      "code": "ステータス",
-      "label": "ステータス",
+      "code": "Status",
+      "label": "Status",
       "enabled": false
     },
     "richText": {
       "type": "RICH_TEXT",
       "code": "richText",
-      "label": "リッチエディター",
+      "label": "richText",
       "noLabel": false,
       "required": false,
       "defaultValue": ""
@@ -66,7 +66,7 @@
     "singleLineText": {
       "type": "SINGLE_LINE_TEXT",
       "code": "singleLineText",
-      "label": "文字列（１行）",
+      "label": "singleLineText",
       "noLabel": false,
       "required": false,
       "minLength": "",
@@ -76,16 +76,16 @@
       "unique": false,
       "defaultValue": ""
     },
-    "カテゴリー": {
+    "Categories": {
       "type": "CATEGORY",
-      "code": "カテゴリー",
-      "label": "カテゴリー",
+      "code": "Categories",
+      "label": "Categories",
       "enabled": false
     },
     "number": {
       "type": "NUMBER",
       "code": "number",
-      "label": "数値",
+      "label": "number",
       "noLabel": false,
       "required": false,
       "minValue": "",
@@ -100,7 +100,7 @@
     "radioButton": {
       "type": "RADIO_BUTTON",
       "code": "radioButton",
-      "label": "ラジオボタン",
+      "label": "radioButton",
       "noLabel": false,
       "required": true,
       "options": {
@@ -119,7 +119,7 @@
     "multiLineText": {
       "type": "MULTI_LINE_TEXT",
       "code": "multiLineText",
-      "label": "文字列 (複数行)",
+      "label": "multiLineText",
       "noLabel": false,
       "required": false,
       "defaultValue": ""
@@ -127,13 +127,13 @@
     "createdTime": {
       "type": "CREATED_TIME",
       "code": "createdTime",
-      "label": "作成日時",
+      "label": "createdTime",
       "noLabel": false
     },
     "checkBox": {
       "type": "CHECK_BOX",
       "code": "checkBox",
-      "label": "チェックボックス",
+      "label": "checkBox",
       "noLabel": false,
       "required": false,
       "options": {
@@ -152,7 +152,7 @@
     "calc": {
       "type": "CALC",
       "code": "calc",
-      "label": "計算",
+      "label": "calc",
       "noLabel": false,
       "required": false,
       "expression": "number + number",
@@ -165,7 +165,7 @@
     "multiSelect": {
       "type": "MULTI_SELECT",
       "code": "multiSelect",
-      "label": "複数選択",
+      "label": "multiSelect",
       "noLabel": false,
       "required": false,
       "options": {
